Preserve caller-provided Number precision in SPC constructors
Pass Number parameters directly to NumFactory instead of converting
through doubleValue(), which drops BigDecimal precision for
DecimalNum-backed series and can shift CUSUM statistics and
control-limit decisions.

+++ ta4j-core/src/main/java/org/ta4j/core/indicators/statistics/CusumIndicator.java
@@ -93,10 +93,10 @@ public class CusumIndicator extends RecursiveCachedIndicator<Num> {
         super(parameters.indicator());
         this.indicator = parameters.indicator();
         NumFactory factory = parameters.indicator().getBarSeries().numFactory();
-        this.targetMean = factory.numOf(parameters.targetMean().doubleValue());
-        this.allowance = factory.numOf(parameters.allowance().doubleValue());
-        this.outlierClipFactor = factory.numOf(parameters.outlierClipFactor().doubleValue());
-        this.scaleDecay = factory.numOf(parameters.scaleDecay().doubleValue());
+        this.targetMean = factory.numOf(parameters.targetMean());
+        this.allowance = factory.numOf(parameters.allowance());
+        this.outlierClipFactor = factory.numOf(parameters.outlierClipFactor());
+        this.scaleDecay = factory.numOf(parameters.scaleDecay());
         this.deviationScale = new DeviationScaleIndicator(this.indicator, this.targetMean, this.allowance,
                 this.scaleDecay);
     }
